Bug 878014 - Use double precision when sampling the delay AudioParam; r=roc
UltraBlame original commit: 9bcaba8046eeab95bd4552a2f68767b4f524a22e

+++ content/media/webaudio/DelayNode.cpp
@@ -173,7 +173,7 @@ public:
     AllocateAudioBlock(numChannels, aOutput);
 
     double delayTime = 0;
-    float computedDelay[WEBAUDIO_BLOCK_SIZE];
+    double computedDelay[WEBAUDIO_BLOCK_SIZE];
     
     const double smoothingRate = WebAudioUtils::ComputeSmoothingRate(0.02, aStream->SampleRate());
